+++ Tests/Test_RadialSolver/Test_PropMatrix/test_a_solid_matrix.py
@@ -5,7 +5,7 @@ import pytest
 import numpy as np
 
 import TidalPy
-TidalPy.test_mode = True
+TidalPy.test_mode()
 
 
 from TidalPy.constants import G