+++ Tests/Test_RadialSolver/Test_PropMatrix/test_b_radial_solver_matrix.py
@@ -2,7 +2,7 @@ import pytest
 import numpy as np
 
 import TidalPy
-TidalPy.test_mode = True
+TidalPy.test_mode()
 
 
 from TidalPy.RadialSolver import radial_solver