+++ Tests/Test_RadialSolver/test_a_boundary_conditions.py
@@ -4,7 +4,7 @@ from math import isnan
 import numpy as np
 
 import TidalPy
-TidalPy.test_mode = True
+TidalPy.test_mode()
 
 from TidalPy.RadialSolver.boundaries.surface_bc import get_surface_bc