+++ Tests/Test_RadialSolver/test_a_starting_condtions.py
@@ -3,7 +3,7 @@ import pytest
 import numpy as np
 
 import TidalPy
-TidalPy.test_mode = True
+TidalPy.test_mode()
 
 from TidalPy.rheology.models import Maxwell
 from TidalPy.RadialSolver.shooting import find_num_shooting_solutions