+++ Tests/Test_RadialSolver/test_e_radial_solver_3layer.py
@@ -4,7 +4,7 @@ import pytest
 import numpy as np
 
 import TidalPy
-TidalPy.test_mode = True
+TidalPy.test_mode()
 
 
 from TidalPy.RadialSolver import radial_solver