+++ Tests/Test_Utilities/Test_Dimensions/test_nondimensional.py
@@ -3,7 +3,7 @@ import pytest
 import numpy as np
 
 import TidalPy
-TidalPy.test_mode = True
+TidalPy.test_mode()
 
 from math import isnan, isclose, sqrt